Refactor: 글로벌 공통 코드 관리를 위한 global 폴더 추가
