+++ src/main/java/com/jerry/ticketing/application/payment/PaymentService.java
@@ -6,9 +6,9 @@ import com.jerry.ticketing.domain.payment.enums.PaymentStatus;
 import com.jerry.ticketing.domain.reservation.Reservation;
 import com.jerry.ticketing.dto.ConfirmTossPayment;
 import com.jerry.ticketing.dto.CreatePayment;
-import com.jerry.ticketing.exception.BusinessException;
-import com.jerry.ticketing.exception.PaymentErrorCode;
-import com.jerry.ticketing.exception.ReservationErrorCode;
+import com.jerry.ticketing.global.exception.BusinessException;
+import com.jerry.ticketing.global.exception.PaymentErrorCode;
+import com.jerry.ticketing.global.exception.ReservationErrorCode;
 import com.jerry.ticketing.repository.payment.PaymentRepository;
 import com.jerry.ticketing.repository.reservation.ReservationRepository;
 import lombok.RequiredArgsConstructor;
