+++ src/main/java/com/jerry/ticketing/global/exception/ConcertErrorCode.java
@@ -1,4 +1,4 @@
-package com.jerry.ticketing.exception;
+package com.jerry.ticketing.global.exception;
 
 
 public enum ConcertErrorCode implements ErrorCode {
